Add core unicode functionality.
Add utf-8 processing routines.
Change ecma_char_t from char/uint16_t to uint16_t.
Apply all utf-8 processing routines.
Change char to jerry_api_char in API functions' declarations.

JerryScript-DCO-1.0-Signed-off-by: Andrey Shitov a.shitov@samsung.com

+++ jerry-core/ecma/builtin-objects/ecma-builtins.cpp
@@ -609,7 +609,7 @@ ecma_builtin_bin_search_for_magic_string_id_in_array (const lit_magic_string_id_
 
     if (ids[mid] == key)
     {
-      return mid;
+      return (int32_t) mid;
     }
     else if (ids[mid] > key)
     {
